The goal here is to utilize the sprockets tmp/cache/assets cache.
Rails3 never used this cache since turbosprockets always deleted
it when doing a precompile.

So, we create the .parcels_sprockets_workaround directory early
enough so that hike's immutable cache is initially created
with a currect 'stat' entry for files inside this directory.
If this doesn't happen, hike's immutable cache writes an entry
with a 'nil' stat for fortitude files that blows up the second
time you attempt to precompile assets.

Also, we need to ensure that fortitude files are ignored during
initial compilation step. The static compiler doesn't exist
in Rails4 anymore, specifically the method each_logical_path was
moved to base.rb. This behavior was previously handled
inside sprockets.rb monkey patch of ::Sprockets::StaticCompiler
and is now handled by the monkey patch to ::Sprockets::Base.