Fix race condition checking workaround directory
With the old code, it was possible for the workaround directory to not
exist when checking if it is a directory, then have been created by
another process by the time it is checked for in order to raise the
error message.

Since FileUtils.mkdir_p will raise an error if attempting to create a 
directory with the same name as a file but will succeed if a directory
with the same name exists, the code can be simplified to eliminate the
race condition.